requirements.txt

Flask
tensorflow>=2.0.0
fastapi
uvicorn
numpy
pillow
pandas
tqdm
